Replace int with NSInteger.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@30608 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOMultiReaderLock.m
@@ -121,8 +121,7 @@
 - (void)lockForReading
 {
   NSThread *ct = [NSThread currentThread];
-
-  int cnt = (int)NSMapGet(_readerThreads,ct);
+  NSInteger cnt = (NSInteger)NSMapGet(_readerThreads,ct);
 
   if (ct == _writerLockThread)
     {
@@ -163,7 +162,8 @@
 - (void)unlockForReading
 {
   NSThread *ct = [NSThread currentThread];
-  int cnt = (int)NSMapGet(_readerThreads,ct);
+  NSInteger cnt = (NSInteger)NSMapGet(_readerThreads,ct);
+
   if (--cnt)
     {
       NSMapInsert(_readerThreads,ct,(void *)(cnt));
